Mus game genera todas las acciones de descarte.
Además, Strategy vuelve a devolver un array de f64 para las
probabilidades.

+++ src/mus/arena/agent.rs
@@ -67,7 +67,7 @@ impl AgenteMusolver {
         }
     }
 
-    fn accion_aleatoria(actions: &[Accion], probabilities: &[u8]) -> Accion {
+    fn accion_aleatoria(actions: &[Accion], probabilities: &[f64]) -> Accion {
         let dist = WeightedIndex::new(probabilities).unwrap();
         let idx = dist.sample(&mut rand::thread_rng());
         actions[idx]
